feat: add support for all upstream dist to file package
use specific function for unix goos and fallback for non-unix systems
